requirements.txt
beautifulsoup4==4.6.0
dj-database-url==0.3.0
Django==1.9.1
django-bootstrap-toolkit==2.15.0
django-bootstrap3==9.0.0
django-extensions==1.9.1
django-geoposition==0.3.0
django-multiselectfield==0.1.8
django-registration-redux==1.3
django-webtest==1.9.2
hellowebapp-deploy==1.0.2
six==1.11.0
waitress==0.8.10
WebOb==1.7.3
WebTest==2.0.28
whitenoise==2.0.6
